Refactoring package ‘report’
‘MetaReport’ meta class has been deleted. Composers functions have been moved in the ‘Report’ class. Package ‘report’ has been deleted because it was containing only the ‘__init__’ module after refactoring. The ‘report.__init__’ module has been renamed ‘report’ and move as a top-level module of ‘linnote’.
